Time: 61 ms (79.60%), Space: 13.1 MB (90.09%) - LeetHub

+++ 18-4sum/18-4sum.cpp
@@ -7,10 +7,10 @@ public:
         sort(num.begin(),num.end());
         for (int i = 0; i < size; i++) 
         {
-            int target_3 = target - num[i];
+            long target_3 = target - num[i];
             for (int j = i + 1; j < size; j++) 
             {
-                int target_2 = target_3 - num[j];
+                long target_2 = target_3 - num[j];
                 int front = j + 1;
                 int back = size - 1;
                 while(front < back)
